Rendering rewrite working
Text rendering has not been ported yet

+++ Source/Platform/SDL2/SDLTiming.cpp
@@ -4,10 +4,10 @@
 
 float SDLTiming::GetTime()
 {
-	return (float)SDL_GetTicks() / 1000.0;
+	return (float)SDL_GetTicks() / 1000.0f;
 }
 
 void SDLTiming::Sleep(float seconds)
 {
-	SDL_Delay(seconds * 1000.0);
+	SDL_Delay((unsigned int)(seconds * 1000.0f));
 }
